+++ halo2-ecc/src/bls12_381/hash_to_curve.rs
@@ -1,188 +1,25 @@
-//! The chip that implements `draft-irtf-cfrg-hash-to-curve-16`
+//! The chip that implements `draft-irtf-cfrg-hash-to-curve-16` for BLS12-381 (G2).
 //! https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-hash-to-curve-16
-
-use std::marker::PhantomData;
 
 use crate::bigint::utils::decode_into_bn;
 use crate::ecc::hash_to_curve::{
     ExpandMessageChip, HashCurveExt, HashInstructions, HashToCurveInstructions,
 };
-use crate::ecc::EcPoint;
 use crate::ff::Field;
-use crate::fields::FieldChipExt;
-use crate::halo2_base::{AssignedValue, Context, QuantumCell};
+use crate::halo2_base::{Context, QuantumCell};
 use crate::halo2_proofs::plonk::Error;
 use crate::{
     ecc::EccChip,
-    fields::{vector::FieldVector, FieldChip, Selectable},
+    fields::{vector::FieldVector, FieldChip},
 };
 use halo2_base::gates::flex_gate::threads::ThreadManager;
-use halo2_base::gates::{GateInstructions, RangeInstructions};
+use halo2_base::gates::RangeInstructions;
 use halo2_base::halo2_proofs::halo2curves::bls12_381::{Fq2, G2};
 use halo2_base::utils::BigPrimeField;
 use itertools::Itertools;
 use num_bigint::BigUint;
 
 use super::{Fp2Chip, Fp2Point, G2Point};
-
-#[derive(Debug)]
-pub struct HashToCurveChip<
-    'chip,
-    F: BigPrimeField,
-    FC: FieldChip<F>,
-    HC: HashInstructions<F>,
-    C: HashCurveExt,
-> {
-    hash_chip: &'chip HC,
-    ecc_chip: EccChip<'chip, F, FC>,
-    _curve: PhantomData<C>,
-}
-
-impl<
-        'chip,
-        F: BigPrimeField,
-        C: HashCurveExt<Base = FC::FieldType>,
-        FC: FieldChipExt<F>,
-        HC: HashInstructions<F> + 'chip,
-    > HashToCurveChip<'chip, F, FC, HC, C>
-where
-    FC::FieldType: crate::ff::PrimeField,
-    FC: Selectable<F, FC::FieldPoint>,
-    EccChip<'chip, F, FC>: HashToCurveInstructions<F, FC, C>,
-{
-    pub fn new(hash_chip: &'chip HC, field_chip: &'chip FC) -> Self {
-        Self { hash_chip, ecc_chip: EccChip::new(field_chip), _curve: PhantomData }
-    }
-
-    pub fn hash_to_curve<XC: ExpandMessageChip>(
-        &self,
-        thread_pool: &mut HC::ThreadManager,
-        msg: impl Iterator<Item = QuantumCell<F>>,
-        dst: &[u8],
-    ) -> Result<EcPoint<F, FC::FieldPoint>, Error> {
-        let u = self.ecc_chip.hash_to_field::<_, XC>(thread_pool, self.hash_chip, msg, dst)?;
-        let p = self.map_to_curve(thread_pool.main(), u)?;
-        Ok(p)
-    }
-
-    fn map_to_curve(
-        &self,
-        ctx: &mut Context<F>,
-        u: [FC::FieldPoint; 2],
-    ) -> Result<EcPoint<F, FC::FieldPoint>, Error> {
-        let [u0, u1] = u;
-
-        let p1 = self.map_to_curve_simple_swu(ctx, u0);
-        let p2 = self.map_to_curve_simple_swu(ctx, u1);
-
-        let p_sum = self.ecc_chip.add_unequal(ctx, p1, p2, false);
-
-        let iso_p = self.ecc_chip.isogeny_map(ctx, p_sum);
-
-        Ok(self.ecc_chip.clear_cofactor(ctx, iso_p))
-    }
-
-    /// Implements [section 6.2 of draft-irtf-cfrg-hash-to-curve-16][map_to_curve_simple_swu]
-    ///
-    /// [map_to_curve_simple_swu]: https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-hash-to-curve-16#appendix-F.2
-    ///
-    /// References:
-    /// - https://github.com/mikelodder7/bls12_381_plus/blob/ml/0.5.6/src/hash_to_curve/map_g2.rs#L388
-    /// - https://github.com/paulmillr/noble-curves/blob/bf70ba9/src/abstract/weierstrass.ts#L1175
-    fn map_to_curve_simple_swu(
-        &self,
-        ctx: &mut Context<F>,
-        u: FC::FieldPoint,
-    ) -> EcPoint<F, FC::FieldPoint> {
-        let field_chip = self.ecc_chip.field_chip();
-        let gate = field_chip.range().gate();
-
-        // constants
-        let swu_a = field_chip.load_constant(ctx, C::SWU_A);
-        let swu_b = field_chip.load_constant(ctx, C::SWU_B);
-        let swu_z = field_chip.load_constant(ctx, C::SWU_Z);
-        let fq2_one = field_chip.load_constant(ctx, <C::Base as crate::ff::Field>::ONE);
-
-        let usq = field_chip.mul(ctx, u.clone(), u.clone()); // 1.  tv1 = u^2
-        let z_usq = field_chip.mul(ctx, usq, swu_z.clone()); // 2.  tv1 = Z * tv1
-        let zsq_u4 = field_chip.mul(ctx, z_usq.clone(), z_usq.clone()); // 3.  tv2 = tv1^2
-        let tv2 = field_chip.add(ctx, zsq_u4, z_usq.clone()); // 4.  tv2 = tv2 + tv1
-        let tv3 = field_chip.add_no_carry(ctx, tv2.clone(), fq2_one); // 5.  tv3 = tv2 + 1
-        let x0_num = field_chip.mul(ctx, tv3, swu_b.clone()); // 6.  tv3 = B * tv3
-
-        let x_den = {
-            let tv2_is_zero = field_chip.is_zero(ctx, tv2.clone());
-            let tv2_neg = field_chip.negate(ctx, tv2);
-
-            field_chip.select(ctx, swu_z, tv2_neg, tv2_is_zero) // tv2_is_zero ? swu_z : tv2_neg
-        }; // 7.  tv4 = tv2 != 0 ? -tv2 : Z
-
-        let x_den = field_chip.mul(ctx, x_den, swu_a.clone()); // 8.  tv4 = A * tv4
-
-        let x0_num_sqr = field_chip.mul(ctx, x0_num.clone(), x0_num.clone()); // 9.  tv2 = tv3^2
-        let x_densq = field_chip.mul(ctx, x_den.clone(), x_den.clone()); // 10. tv6 = tv4^2
-        let ax_densq = field_chip.mul(ctx, x_densq.clone(), swu_a); // 11. tv5 = A * tv6
-        let tv2 = field_chip.add_no_carry(ctx, x0_num_sqr, ax_densq); // 12. tv2 = tv2 + tv5
-        let tv2 = field_chip.mul(ctx, tv2, x0_num.clone()); // 13. tv2 = tv2 * tv3
-        let gx_den = field_chip.mul(ctx, x_densq, x_den.clone()); // 14. tv6 = tv6 * tv4
-        let tv5 = field_chip.mul(ctx, gx_den.clone(), swu_b); // 15. tv5 = B * tv6
-        let gx0_num = field_chip.add(ctx, tv2, tv5); // 16. tv2 = tv2 + tv5
-
-        let x = field_chip.mul(ctx, &z_usq, &x0_num); // 17.  x = tv1 * tv3
-
-        let (is_gx1_square, y1) = self.sqrt_ratio(ctx, gx0_num, gx_den); // 18.  (is_gx1_square, y1) = sqrt_ratio(tv2, tv6)
-
-        let y = field_chip.mul(ctx, &z_usq, &u); // 19.  y = tv1 * u
-        let y = field_chip.mul(ctx, y, y1.clone()); // 20.  y = y * y1
-        let x = field_chip.select(ctx, x0_num, x, is_gx1_square); // 21.  x = is_gx1_square ? tv3 : x
-        let y = field_chip.select(ctx, y1, y, is_gx1_square); // 22.  y = is_gx1_square ? y1 : y
-
-        let to_neg = {
-            let u_sgn = field_chip.sgn0(ctx, u);
-            let y_sgn = field_chip.sgn0(ctx, y.clone());
-            gate.xor(ctx, u_sgn, y_sgn)
-        }; // 23.  e1 = sgn0(u) == sgn0(y) // we implement an opposite condition: !e1 = sgn0(u) ^ sgn0(y)
-
-        let y_neg = field_chip.negate(ctx, y.clone());
-        let y = field_chip.select(ctx, y_neg, y, to_neg); // 24.  y = !e1 ? -y : y
-        let x = field_chip.divide(ctx, x, x_den); // 25.  x = x / tv4
-
-        EcPoint::new(x, y)
-    }
-
-    // Implements [Appendix F.2.1 of draft-irtf-cfrg-hash-to-curve-16][sqrt_ration]
-    //
-    // [sqrt_ration]: https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-hash-to-curve-16#appendix-F.2.1
-    fn sqrt_ratio(
-        &self,
-        ctx: &mut Context<F>,
-        num: FC::FieldPoint,
-        div: FC::FieldPoint,
-    ) -> (AssignedValue<F>, FC::FieldPoint) {
-        let field_chip = self.ecc_chip.field_chip();
-        let num_v = field_chip.get_assigned_value(&num.clone().into());
-        let div_v = field_chip.get_assigned_value(&div.clone().into());
-
-        let (is_square, y) = FC::FieldType::sqrt_ratio(&num_v, &div_v);
-
-        let is_square = ctx.load_witness(F::from(is_square.unwrap_u8() as u64));
-        field_chip.gate().assert_bit(ctx, is_square); // assert is_square is boolean
-
-        let y_assigned = field_chip.load_private(ctx, y);
-        let y_sqr = field_chip.mul(ctx, y_assigned.clone(), y_assigned.clone()); // y_sqr = y1^2
-
-        let ratio = field_chip.divide(ctx, num, div); // r = u / v
-
-        let swu_z = field_chip.load_constant(ctx, C::SWU_Z);
-        let ratio_z = field_chip.mul(ctx, ratio.clone(), swu_z.clone()); // r_z = r * z
-
-        let y_check = field_chip.select(ctx, ratio, ratio_z, is_square); // y_check = is_square ? ratio : r_z
-
-        field_chip.assert_equal(ctx, y_check, y_sqr); // assert y_check == y_sqr
-
-        (is_square, y_assigned)
-    }
-}
 
 const G2_EXT_DEGREE: usize = 2;
 // L = ceil((ceil(log2(p)) + k) / 8) (see section 5 of ietf draft link above)
@@ -349,8 +186,6 @@ mod bls12_381 {
     use super::HashCurveExt;
 
     impl HashCurveExt for G2 {
-        type Fp = Fq;
-
         const BLS_X: u64 = 0xd201000000010000;
 
         const SWU_A: Self::Base = Self::Base {
