+++ halo2-ecc/src/ecc/hash_to_curve.rs
@@ -1,22 +1,31 @@
-use std::iter;
+//! Implements `draft-irtf-cfrg-hash-to-curve-16`.
+//! https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-hash-to-curve-16
+
+use std::{iter, marker::PhantomData};
 
 use halo2_base::{
-    gates::{GateInstructions, RangeInstructions, flex_gate::threads::ThreadManager},
+    gates::{flex_gate::threads::ThreadManager, GateInstructions, RangeInstructions},
     halo2_proofs::{halo2curves::CurveExt, plonk::Error},
     utils::BigPrimeField,
     AssignedValue, Context, QuantumCell,
 };
 use itertools::Itertools;
 
-use crate::fields::{FieldChipExt, Selectable};
+use crate::ff::Field;
+use crate::fields::{FieldChip, FieldChipExt, Selectable};
 
-use super::{scalar_multiply_bits, EcPoint};
+use super::{scalar_multiply_bits, EcPoint, EccChip};
 
+/// Trait that defines basic interfaces of hash chips supporting custom region manangers (e.g. [`SinglePhaseCoreManager`]).
 pub trait HashInstructions<F: BigPrimeField> {
+    // Number of bytes absorbed in a single rount of hashing.
     const BLOCK_SIZE: usize;
+    // Number of bytes in the output.
     const DIGEST_SIZE: usize;
 
+    // Type of region manager used by the hash function.
     type ThreadManager: ThreadManager<F>;
+    // Type of output produced by the hash function.
     type Output: IntoIterator<Item = AssignedValue<F>>;
 
     /// Digests input using hash function and returns finilized output.
@@ -30,14 +39,11 @@ pub trait HashInstructions<F: BigPrimeField> {
     ) -> Result<Self::Output, Error>;
 }
 
+/// Trait that extneds [`CurveExt`] with constants specific to hash to curve operations.
 pub trait HashCurveExt: CurveExt
 where
     Self::Base: crate::ff::PrimeField,
 {
-    type Fp: crate::ff::PrimeField + crate::ff::WithSmallOrderMulGroup<3>;
-
-    const BLS_X: u64;
-
     const SWU_A: Self::Base;
     const SWU_B: Self::Base;
     const SWU_Z: Self::Base;
@@ -50,11 +56,14 @@ where
     const PSI_X: Self::Base;
     const PSI_Y: Self::Base;
     const PSI2_X: Self::Base;
+
+    const BLS_X: u64;
 }
 
+/// A trait for message expansion methods supported by [`HashToCurveChip`].
 pub trait ExpandMessageChip {
     fn expand_message<F: BigPrimeField, HC: HashInstructions<F>>(
-        ctx: &mut HC::ThreadManager,
+        thread_pool: &mut HC::ThreadManager,
         hash_chip: &HC,
         range: &impl RangeInstructions<F>,
         msg: impl Iterator<Item = QuantumCell<F>>,
@@ -63,6 +72,7 @@ pub trait ExpandMessageChip {
     ) -> Result<Vec<AssignedValue<F>>, Error>;
 }
 
+/// Trait that defines methods used by the [`HashToCurveChip`] to the curve point for [`EccChip`].
 pub trait HashToCurveInstructions<
     F: BigPrimeField,
     FC: FieldChipExt<F>,
@@ -73,6 +83,7 @@ pub trait HashToCurveInstructions<
 {
     fn field_chip(&self) -> &FC;
 
+    /// Computes `scalar * P` where scalar is represented as vector of [bits].
     fn scalar_mult_bits(
         &self,
         ctx: &mut Context<F>,
@@ -84,6 +95,12 @@ pub trait HashToCurveInstructions<
         scalar_multiply_bits(self.field_chip(), ctx, p, bits, max_bits, window_bits, true)
     }
 
+    /// Hashes a byte string of arbitrary length into one or more elements of `Self`,
+    /// using [`ExpandMessage`] variant `X`.
+    ///
+    /// Implements [section 5.2 of `draft-irtf-cfrg-hash-to-curve-16`][hash_to_field].
+    ///
+    /// [hash_to_field]: https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-hash-to-curve-16#section-5.2
     fn hash_to_field<HC: HashInstructions<F>, XC: ExpandMessageChip>(
         &self,
         thread_pool: &mut HC::ThreadManager,
@@ -92,18 +109,25 @@ pub trait HashToCurveInstructions<
         dst: &[u8],
     ) -> Result<[FC::FieldPoint; 2], Error>;
 
+    /// Implements [Appendix E.3 of draft-irtf-cfrg-hash-to-curve-16][isogeny_map_g2]
+    ///
+    /// [isogeny_map_g2]: https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-hash-to-curve-16#appendix-E.3
     fn isogeny_map(
         &self,
         ctx: &mut Context<F>,
         p: EcPoint<F, FC::FieldPoint>,
     ) -> EcPoint<F, FC::FieldPoint>;
 
+    /// Implements [Appendix G.3 of draft-irtf-cfrg-hash-to-curve-16][clear_cofactor]
+    ///
+    /// [clear_cofactor]: https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-hash-to-curve-16#appendix-G.3
     fn clear_cofactor(
         &self,
         ctx: &mut Context<F>,
         p: EcPoint<F, FC::FieldPoint>,
     ) -> EcPoint<F, FC::FieldPoint>;
 
+    /// Computes `scalar * P` where scalar is [C::BLS_X] variant of [C: HashCurveExt].
     fn mul_by_bls_x(
         &self,
         ctx: &mut Context<F>,
@@ -117,6 +141,7 @@ pub trait HashToCurveInstructions<
         self.scalar_mult_bits(ctx, p, bls_x_bits, 4)
     }
 
+    /// Computes the endomorphism psi for point [p].
     fn psi(
         &self,
         ctx: &mut Context<F>,
@@ -137,6 +162,7 @@ pub trait HashToCurveInstructions<
         EcPoint::new(x, y)
     }
 
+    /// Efficiently omputes psi(psi(P)) for point [p].
     fn psi2(
         &self,
         ctx: &mut Context<F>,
@@ -152,6 +178,174 @@ pub trait HashToCurveInstructions<
     }
 }
 
+/// Implementation of random oracle maps to the curve.
+#[derive(Debug)]
+pub struct HashToCurveChip<
+    'chip,
+    F: BigPrimeField,
+    FC: FieldChip<F>,
+    HC: HashInstructions<F>,
+    C: HashCurveExt,
+> {
+    hash_chip: &'chip HC,
+    ecc_chip: EccChip<'chip, F, FC>,
+    _curve: PhantomData<C>,
+}
+
+impl<
+        'chip,
+        F: BigPrimeField,
+        C: HashCurveExt<Base = FC::FieldType>,
+        FC: FieldChipExt<F>,
+        HC: HashInstructions<F> + 'chip,
+    > HashToCurveChip<'chip, F, FC, HC, C>
+where
+    FC::FieldType: crate::ff::PrimeField,
+    FC: Selectable<F, FC::FieldPoint>,
+    EccChip<'chip, F, FC>: HashToCurveInstructions<F, FC, C>,
+{
+    pub fn new(hash_chip: &'chip HC, field_chip: &'chip FC) -> Self {
+        Self { hash_chip, ecc_chip: EccChip::new(field_chip), _curve: PhantomData }
+    }
+
+    /// Implements a uniform encoding from byte strings to elements of [`EcPoint<F, FC::FieldPoint>`].
+    pub fn hash_to_curve<XC: ExpandMessageChip>(
+        &self,
+        thread_pool: &mut HC::ThreadManager,
+        msg: impl Iterator<Item = QuantumCell<F>>,
+        dst: &[u8],
+    ) -> Result<EcPoint<F, FC::FieldPoint>, Error> {
+        let u = self.ecc_chip.hash_to_field::<_, XC>(thread_pool, self.hash_chip, msg, dst)?;
+        let p = self.map_to_curve(thread_pool.main(), u)?;
+        Ok(p)
+    }
+
+    /// Maps an element of the finite field `FC::FieldPoint` to a point on the curve [`EcPoint<F, FC::FieldPoint>`].
+    fn map_to_curve(
+        &self,
+        ctx: &mut Context<F>,
+        u: [FC::FieldPoint; 2],
+    ) -> Result<EcPoint<F, FC::FieldPoint>, Error> {
+        let [u0, u1] = u;
+
+        let p1 = self.map_to_curve_simple_swu(ctx, u0);
+        let p2 = self.map_to_curve_simple_swu(ctx, u1);
+
+        let p_sum = self.ecc_chip.add_unequal(ctx, p1, p2, false);
+
+        let iso_p = self.ecc_chip.isogeny_map(ctx, p_sum);
+
+        Ok(self.ecc_chip.clear_cofactor(ctx, iso_p))
+    }
+
+    /// Implements [section 6.2 of draft-irtf-cfrg-hash-to-curve-16][map_to_curve_simple_swu]
+    ///
+    /// [map_to_curve_simple_swu]: https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-hash-to-curve-16#appendix-F.2
+    ///
+    /// References:
+    /// - https://github.com/mikelodder7/bls12_381_plus/blob/ml/0.5.6/src/hash_to_curve/map_g2.rs#L388
+    /// - https://github.com/paulmillr/noble-curves/blob/bf70ba9/src/abstract/weierstrass.ts#L1175
+    fn map_to_curve_simple_swu(
+        &self,
+        ctx: &mut Context<F>,
+        u: FC::FieldPoint,
+    ) -> EcPoint<F, FC::FieldPoint> {
+        let field_chip = self.ecc_chip.field_chip();
+        let gate = field_chip.range().gate();
+
+        // constants
+        let swu_a = field_chip.load_constant(ctx, C::SWU_A);
+        let swu_b = field_chip.load_constant(ctx, C::SWU_B);
+        let swu_z = field_chip.load_constant(ctx, C::SWU_Z);
+        let fq2_one = field_chip.load_constant(ctx, <C::Base as crate::ff::Field>::ONE);
+
+        let usq = field_chip.mul(ctx, u.clone(), u.clone()); // 1.  tv1 = u^2
+        let z_usq = field_chip.mul(ctx, usq, swu_z.clone()); // 2.  tv1 = Z * tv1
+        let zsq_u4 = field_chip.mul(ctx, z_usq.clone(), z_usq.clone()); // 3.  tv2 = tv1^2
+        let tv2 = field_chip.add(ctx, zsq_u4, z_usq.clone()); // 4.  tv2 = tv2 + tv1
+        let tv3 = field_chip.add_no_carry(ctx, tv2.clone(), fq2_one); // 5.  tv3 = tv2 + 1
+        let x0_num = field_chip.mul(ctx, tv3, swu_b.clone()); // 6.  tv3 = B * tv3
+
+        let x_den = {
+            let tv2_is_zero = field_chip.is_zero(ctx, tv2.clone());
+            let tv2_neg = field_chip.negate(ctx, tv2);
+
+            field_chip.select(ctx, swu_z, tv2_neg, tv2_is_zero) // tv2_is_zero ? swu_z : tv2_neg
+        }; // 7.  tv4 = tv2 != 0 ? -tv2 : Z
+
+        let x_den = field_chip.mul(ctx, x_den, swu_a.clone()); // 8.  tv4 = A * tv4
+
+        let x0_num_sqr = field_chip.mul(ctx, x0_num.clone(), x0_num.clone()); // 9.  tv2 = tv3^2
+        let x_densq = field_chip.mul(ctx, x_den.clone(), x_den.clone()); // 10. tv6 = tv4^2
+        let ax_densq = field_chip.mul(ctx, x_densq.clone(), swu_a); // 11. tv5 = A * tv6
+        let tv2 = field_chip.add_no_carry(ctx, x0_num_sqr, ax_densq); // 12. tv2 = tv2 + tv5
+        let tv2 = field_chip.mul(ctx, tv2, x0_num.clone()); // 13. tv2 = tv2 * tv3
+        let gx_den = field_chip.mul(ctx, x_densq, x_den.clone()); // 14. tv6 = tv6 * tv4
+        let tv5 = field_chip.mul(ctx, gx_den.clone(), swu_b); // 15. tv5 = B * tv6
+        let gx0_num = field_chip.add(ctx, tv2, tv5); // 16. tv2 = tv2 + tv5
+
+        let x = field_chip.mul(ctx, &z_usq, &x0_num); // 17.  x = tv1 * tv3
+
+        let (is_gx1_square, y1) = self.sqrt_ratio(ctx, gx0_num, gx_den); // 18.  (is_gx1_square, y1) = sqrt_ratio(tv2, tv6)
+
+        let y = field_chip.mul(ctx, &z_usq, &u); // 19.  y = tv1 * u
+        let y = field_chip.mul(ctx, y, y1.clone()); // 20.  y = y * y1
+        let x = field_chip.select(ctx, x0_num, x, is_gx1_square); // 21.  x = is_gx1_square ? tv3 : x
+        let y = field_chip.select(ctx, y1, y, is_gx1_square); // 22.  y = is_gx1_square ? y1 : y
+
+        let to_neg = {
+            let u_sgn = field_chip.sgn0(ctx, u);
+            let y_sgn = field_chip.sgn0(ctx, y.clone());
+            gate.xor(ctx, u_sgn, y_sgn)
+        }; // 23.  e1 = sgn0(u) == sgn0(y) // we implement an opposite condition: !e1 = sgn0(u) ^ sgn0(y)
+
+        let y_neg = field_chip.negate(ctx, y.clone());
+        let y = field_chip.select(ctx, y_neg, y, to_neg); // 24.  y = !e1 ? -y : y
+        let x = field_chip.divide(ctx, x, x_den); // 25.  x = x / tv4
+
+        EcPoint::new(x, y)
+    }
+
+    // Implements [Appendix F.2.1 of draft-irtf-cfrg-hash-to-curve-16][sqrt_ration]
+    //
+    // [sqrt_ration]: https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-hash-to-curve-16#appendix-F.2.1
+    fn sqrt_ratio(
+        &self,
+        ctx: &mut Context<F>,
+        num: FC::FieldPoint,
+        div: FC::FieldPoint,
+    ) -> (AssignedValue<F>, FC::FieldPoint) {
+        let field_chip = self.ecc_chip.field_chip();
+        let num_v = field_chip.get_assigned_value(&num.clone().into());
+        let div_v = field_chip.get_assigned_value(&div.clone().into());
+
+        let (is_square, y) = FC::FieldType::sqrt_ratio(&num_v, &div_v);
+
+        let is_square = ctx.load_witness(F::from(is_square.unwrap_u8() as u64));
+        field_chip.gate().assert_bit(ctx, is_square); // assert is_square is boolean
+
+        let y_assigned = field_chip.load_private(ctx, y);
+        let y_sqr = field_chip.mul(ctx, y_assigned.clone(), y_assigned.clone()); // y_sqr = y1^2
+
+        let ratio = field_chip.divide(ctx, num, div); // r = u / v
+
+        let swu_z = field_chip.load_constant(ctx, C::SWU_Z);
+        let ratio_z = field_chip.mul(ctx, ratio.clone(), swu_z.clone()); // r_z = r * z
+
+        let y_check = field_chip.select(ctx, ratio, ratio_z, is_square); // y_check = is_square ? ratio : r_z
+
+        field_chip.assert_equal(ctx, y_check, y_sqr); // assert y_check == y_sqr
+
+        (is_square, y_assigned)
+    }
+}
+
+/// Constructor for `expand_message_xmd` for a given digest hash function, message, DST,
+/// and output length.
+///
+/// Implements [section 5.3.1 of `draft-irtf-cfrg-hash-to-curve-16`][expand_message_xmd].
+///
+/// [expand_message_xmd]: https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-hash-to-curve-16#section-5.3.1
 pub struct ExpandMsgXmd;
 
 impl ExpandMessageChip for ExpandMsgXmd {
@@ -228,13 +422,21 @@ impl ExpandMessageChip for ExpandMsgXmd {
         );
 
         for i in 1..ell {
-            let preimg = strxor(b_0.iter().copied(), b_vals[i - 1].iter().copied(), gate, thread_pool.main())
-                .into_iter()
-                .chain(iter::once(thread_pool.main().load_constant(F::from(i as u64 + 1))))
-                .chain(dst_prime.clone())
-                .map(QuantumCell::Existing);
+            let preimg = strxor(
+                b_0.iter().copied(),
+                b_vals[i - 1].iter().copied(),
+                gate,
+                thread_pool.main(),
+            )
+            .into_iter()
+            .chain(iter::once(thread_pool.main().load_constant(F::from(i as u64 + 1))))
+            .chain(dst_prime.clone())
+            .map(QuantumCell::Existing);
 
-            b_vals.insert(i, hash_chip.digest::<77>(thread_pool, preimg, false)?.into_iter().collect_vec());
+            b_vals.insert(
+                i,
+                hash_chip.digest::<77>(thread_pool, preimg, false)?.into_iter().collect_vec(),
+            );
         }
 
         let uniform_bytes = b_vals.into_iter().flatten().take(len_in_bytes).collect_vec();
